tools/requirements.txt
sentence-transformers>=2.0.0
pyyaml>=6.0
